+++ familyhelper-plugin-life/src/main/java/com/dwarfeng/familyhelper/plugin/life/bean/dto/DubboRestActivityFileStream.java
@@ -1,0 +1,73 @@
+package com.dwarfeng.familyhelper.plugin.life.bean.dto;
+
+import com.dwarfeng.subgrade.stack.bean.dto.Dto;
+import org.jboss.resteasy.annotations.jaxrs.FormParam;
+import org.jboss.resteasy.annotations.providers.multipart.PartType;
+
+import javax.ws.rs.core.MediaType;
+import java.io.InputStream;
+
+/**
+ * Dubbo rest 活动文件流。
+ *
+ * @author DwArFeng
+ * @since 1.5.0
+ */
+public class DubboRestActivityFileStream implements Dto {
+
+    private static final long serialVersionUID = 55366170030723843L;
+
+    @FormParam("origin-name")
+    @PartType("text/plain;charset=utf-8")
+    private String originName;
+
+    @FormParam("length")
+    @PartType(MediaType.TEXT_PLAIN)
+    private long length;
+
+    @FormParam("content")
+    @PartType(MediaType.APPLICATION_OCTET_STREAM)
+    private InputStream content;
+
+    public DubboRestActivityFileStream() {
+    }
+
+    public DubboRestActivityFileStream(String originName, long length, InputStream content) {
+        this.originName = originName;
+        this.length = length;
+        this.content = content;
+    }
+
+    public String getOriginName() {
+        return originName;
+    }
+
+    public void setOriginName(String originName) {
+        this.originName = originName;
+    }
+
+    public long getLength() {
+        return length;
+    }
+
+    public void setLength(long length) {
+        this.length = length;
+    }
+
+    public InputStream getContent() {
+        return content;
+    }
+
+    public void setContent(InputStream content) {
+        this.content = content;
+    }
+
+    @Override
+    public String toString() {
+        return "DubboRestActivityFileStream{" +
+                "originName='" + originName + '\'' +
+                ", length=" + length +
+                ", content=" + content +
+                '}';
+    }
+}